+++ src/lib.rs
@@ -42,21 +42,22 @@ pub use crate::traits::GerberCode;
 pub use crate::types::*;
 
 #[cfg(test)]
-mod test {
-    use std::io::BufWriter;
-
+mod serializaion_tests {
     use super::traits::PartialGerberCode;
     use super::*;
+    use chrono::DateTime;
+    use std::io::BufWriter;
+    use uuid::Uuid;
 
     #[test]
-    fn test_serialize() {
+    fn test_comment() {
         //! The serialize method of the GerberCode trait should generate strings.
         let comment = GCode::Comment("testcomment".to_string());
         assert_code!(comment, "G04 testcomment*\n");
     }
 
     #[test]
-    fn test_vec_serialize() {
+    fn test_vec_of_comments() {
         //! A `Vec<T: GerberCode>` should also implement `GerberCode`.
         let mut v = Vec::new();
         v.push(GCode::Comment("comment 1".to_string()));
@@ -65,7 +66,7 @@ mod test {
     }
 
     #[test]
-    fn test_command_serialize() {
+    fn test_single_command() {
         //! A `Command` should implement `GerberCode`
         let c = Command::FunctionCode(FunctionCode::GCode(GCode::Comment("comment".to_string())));
         assert_code!(c, "G04 comment*\n");
@@ -264,7 +265,7 @@ mod test {
     }
 
     #[test]
-    fn test_polarity_serialize() {
+    fn test_polarity() {
         let d = ExtendedCode::LoadPolarity(Polarity::Dark);
         let c = ExtendedCode::LoadPolarity(Polarity::Clear);
         assert_code!(d, "%LPD*%\n");
@@ -272,7 +273,7 @@ mod test {
     }
 
     #[test]
-    fn test_step_and_repeat_serialize() {
+    fn test_step_and_repeat() {
         let o = ExtendedCode::StepAndRepeat(StepAndRepeat::Open {
             repeat_x: 2,
             repeat_y: 3,
@@ -285,29 +286,450 @@ mod test {
     }
 
     #[test]
-    fn test_delete_attribute_serialize() {
+    fn test_delete_attribute() {
         let d = ExtendedCode::DeleteAttribute("foo".into());
         assert_code!(d, "%TDfoo*%\n");
     }
 
     #[test]
-    fn test_file_attribute_serialize() {
-        let part = ExtendedCode::FileAttribute(FileAttribute::Part(Part::Other("foo".into())));
-        assert_code!(part, "%TF.Part,Other,foo*%\n");
-
-        let gensw1 = ExtendedCode::FileAttribute(FileAttribute::GenerationSoftware(
-            GenerationSoftware::new("Vend0r", "superpcb", None),
-        ));
-        assert_code!(gensw1, "%TF.GenerationSoftware,Vend0r,superpcb*%\n");
-
-        let gensw2 = ExtendedCode::FileAttribute(FileAttribute::GenerationSoftware(
-            GenerationSoftware::new("Vend0r", "superpcb", Some("1.2.3")),
-        ));
-        assert_code!(gensw2, "%TF.GenerationSoftware,Vend0r,superpcb,1.2.3*%\n");
+    fn test_file_attribute_part() {
+        let part = ExtendedCode::FileAttribute(FileAttribute::Part(Part::Other("Part 1".into())));
+        assert_code!(part, "%TF.Part,Other,Part 1*%\n");
     }
 
     #[test]
-    fn test_aperture_attribute_serialize() {
+    fn test_file_attribute_generation_software() {
+        let gensw1 = ExtendedCode::FileAttribute(FileAttribute::GenerationSoftware(
+            GenerationSoftware::new("Vendor 1", "App 1", None),
+        ));
+        assert_code!(gensw1, "%TF.GenerationSoftware,Vendor 1,App 1*%\n");
+
+        let gensw2 = ExtendedCode::FileAttribute(FileAttribute::GenerationSoftware(
+            GenerationSoftware::new("Vendor 1", "App 1", Some("1.2.3")),
+        ));
+        assert_code!(gensw2, "%TF.GenerationSoftware,Vendor 1,App 1,1.2.3*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_creation_date() {
+        let date = DateTime::parse_from_rfc3339("2025-06-10T16:25:00+02:00").unwrap();
+        let date = ExtendedCode::FileAttribute(FileAttribute::CreationDate(date));
+        assert_code!(date, "%TF.CreationDate,2025-06-10T16:25:00+02:00*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_project_id() {
+        let proj = ExtendedCode::FileAttribute(FileAttribute::ProjectId {
+            id: "Project".into(),
+            guid: Uuid::max(),
+            revision: "rev1".into(),
+        });
+        assert_code!(
+            proj,
+            "%TF.ProjectId,Project,ffffffff-ffff-ffff-ffff-ffffffffffff,rev1*%\n"
+        );
+    }
+
+    #[test]
+    fn test_file_attribute_file_function_copper() {
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Copper {
+            layer: 1,
+            pos: ExtendedPosition::Top,
+            copper_type: None,
+        }));
+        assert_code!(func, "%TF.FileFunction,Copper,L1,Top*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Copper {
+            layer: 2,
+            pos: ExtendedPosition::Bottom,
+            copper_type: Some(CopperType::Hatched),
+        }));
+        assert_code!(func, "%TF.FileFunction,Copper,L2,Bot,Hatched*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Copper {
+            layer: 3,
+            pos: ExtendedPosition::Bottom,
+            copper_type: Some(CopperType::Mixed),
+        }));
+        assert_code!(func, "%TF.FileFunction,Copper,L3,Bot,Mixed*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Copper {
+            layer: 4,
+            pos: ExtendedPosition::Bottom,
+            copper_type: Some(CopperType::Plane),
+        }));
+        assert_code!(func, "%TF.FileFunction,Copper,L4,Bot,Plane*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Copper {
+            layer: 5,
+            pos: ExtendedPosition::Bottom,
+            copper_type: Some(CopperType::Signal),
+        }));
+        assert_code!(func, "%TF.FileFunction,Copper,L5,Bot,Signal*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_file_function_plated() {
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Plated {
+            from_layer: 1,
+            to_layer: 2,
+            drill: Drill::Blind,
+            label: None,
+        }));
+        assert_code!(func, "%TF.FileFunction,Plated,1,2,Blind*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Plated {
+            from_layer: 1,
+            to_layer: 4,
+            drill: Drill::ThroughHole,
+            label: None,
+        }));
+        assert_code!(func, "%TF.FileFunction,Plated,1,4,PTH*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Plated {
+            from_layer: 2,
+            to_layer: 3,
+            drill: Drill::Buried,
+            label: None,
+        }));
+        assert_code!(func, "%TF.FileFunction,Plated,2,3,Buried*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Plated {
+            from_layer: 1,
+            to_layer: 2,
+            drill: Drill::ThroughHole,
+            label: Some(DrillRouteType::Drill),
+        }));
+        assert_code!(func, "%TF.FileFunction,Plated,1,2,PTH,Drill*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Plated {
+            from_layer: 1,
+            to_layer: 2,
+            drill: Drill::ThroughHole,
+            label: Some(DrillRouteType::Mixed),
+        }));
+        assert_code!(func, "%TF.FileFunction,Plated,1,2,PTH,Mixed*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Plated {
+            from_layer: 1,
+            to_layer: 2,
+            drill: Drill::ThroughHole,
+            label: Some(DrillRouteType::Route),
+        }));
+        assert_code!(func, "%TF.FileFunction,Plated,1,2,PTH,Rout*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_file_function_non_plated() {
+        let func =
+            ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::NonPlated {
+                from_layer: 1,
+                to_layer: 2,
+                drill: Drill::Blind,
+                label: None,
+            }));
+        assert_code!(func, "%TF.FileFunction,NonPlated,1,2,Blind*%\n");
+
+        let func =
+            ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::NonPlated {
+                from_layer: 1,
+                to_layer: 4,
+                drill: Drill::ThroughHole,
+                label: None,
+            }));
+        assert_code!(func, "%TF.FileFunction,NonPlated,1,4,PTH*%\n");
+
+        let func =
+            ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::NonPlated {
+                from_layer: 2,
+                to_layer: 3,
+                drill: Drill::Buried,
+                label: None,
+            }));
+        assert_code!(func, "%TF.FileFunction,NonPlated,2,3,Buried*%\n");
+
+        let func =
+            ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::NonPlated {
+                from_layer: 1,
+                to_layer: 2,
+                drill: Drill::ThroughHole,
+                label: Some(DrillRouteType::Drill),
+            }));
+        assert_code!(func, "%TF.FileFunction,NonPlated,1,2,PTH,Drill*%\n");
+
+        let func =
+            ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::NonPlated {
+                from_layer: 1,
+                to_layer: 2,
+                drill: Drill::ThroughHole,
+                label: Some(DrillRouteType::Mixed),
+            }));
+        assert_code!(func, "%TF.FileFunction,NonPlated,1,2,PTH,Mixed*%\n");
+
+        let func =
+            ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::NonPlated {
+                from_layer: 1,
+                to_layer: 2,
+                drill: Drill::ThroughHole,
+                label: Some(DrillRouteType::Route),
+            }));
+        assert_code!(func, "%TF.FileFunction,NonPlated,1,2,PTH,Rout*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_file_function_profile() {
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Profile(
+            Profile::Plated,
+        )));
+        assert_code!(func, "%TF.FileFunction,Profile,P*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::Profile(
+            Profile::NonPlated,
+        )));
+        assert_code!(func, "%TF.FileFunction,Profile,NP*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_file_function_keepout() {
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::KeepOut(
+            Position::Top,
+        )));
+        assert_code!(func, "%TF.FileFunction,Keepout,Top*%\n");
+
+        let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::KeepOut(
+            Position::Bottom,
+        )));
+        assert_code!(func, "%TF.FileFunction,Keepout,Bot*%\n");
+    }
+
+    macro_rules! test_position_and_index {
+        ($test:ident, $ff:ident, $value:literal) => {
+            #[test]
+            fn $test() {
+                let func =
+                    ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::$ff {
+                        pos: Position::Top,
+                        index: None,
+                    }));
+                assert_code!(func, &format!("%TF.FileFunction,{},Top*%\n", $value));
+
+                let func =
+                    ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::$ff {
+                        pos: Position::Top,
+                        index: Some(1),
+                    }));
+                assert_code!(func, &format!("%TF.FileFunction,{},Top,1*%\n", $value));
+
+                let func =
+                    ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::$ff {
+                        pos: Position::Bottom,
+                        index: None,
+                    }));
+                assert_code!(func, &format!("%TF.FileFunction,{},Bot*%\n", $value));
+            }
+        };
+    }
+
+    macro_rules! test_layer_and_position {
+        ($test:ident, $ff:ident, $value:literal) => {
+            #[test]
+            fn $test() {
+                let func =
+                    ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::$ff {
+                        pos: Position::Top,
+                        layer: 1,
+                    }));
+                assert_code!(func, &format!("%TF.FileFunction,{},L{},Top*%\n", $value, 1));
+
+                let func =
+                    ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::$ff {
+                        pos: Position::Bottom,
+                        layer: 2,
+                    }));
+                assert_code!(func, &format!("%TF.FileFunction,{},L{},Bot*%\n", $value, 2));
+            }
+        };
+    }
+
+    macro_rules! test_position {
+        ($test:ident, $ff:ident, $value:literal) => {
+            #[test]
+            fn $test() {
+                let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(
+                    FileFunction::$ff(Position::Top),
+                ));
+                assert_code!(func, &format!("%TF.FileFunction,{},Top*%\n", $value));
+
+                let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(
+                    FileFunction::$ff(Position::Bottom),
+                ));
+                assert_code!(func, &format!("%TF.FileFunction,{},Bot*%\n", $value));
+            }
+        };
+    }
+
+    macro_rules! test_optional_position {
+        ($test:ident, $ff:ident, $value:literal) => {
+            #[test]
+            fn $test() {
+                let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(
+                    FileFunction::$ff(None),
+                ));
+                assert_code!(func, &format!("%TF.FileFunction,{}*%\n", $value));
+
+                let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(
+                    FileFunction::$ff(Some(Position::Top)),
+                ));
+                assert_code!(func, &format!("%TF.FileFunction,{},Top*%\n", $value));
+
+                let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(
+                    FileFunction::$ff(Some(Position::Bottom)),
+                ));
+                assert_code!(func, &format!("%TF.FileFunction,{},Bot*%\n", $value));
+            }
+        };
+    }
+    macro_rules! test_simple {
+        ($test:ident, $ff:ident, $value:literal) => {
+            #[test]
+            fn $test() {
+                let func =
+                    ExtendedCode::FileAttribute(FileAttribute::FileFunction(FileFunction::$ff));
+                assert_code!(func, &format!("%TF.FileFunction,{}*%\n", $value));
+            }
+        };
+    }
+
+    macro_rules! test_string {
+        ($test:ident, $ff:ident, $value:literal) => {
+            #[test]
+            fn $test() {
+                let string = "A String".to_string();
+                let func = ExtendedCode::FileAttribute(FileAttribute::FileFunction(
+                    FileFunction::$ff(string.clone()),
+                ));
+                assert_code!(func, &format!("%TF.FileFunction,{},{}*%\n", $value, string));
+            }
+        };
+    }
+
+    //
+    // It should be noted that the gerber spec is very-inconsistent with casing, e.g. "Soldermask" vs "AssemblyDrawing"
+    //
+
+    test_position_and_index!(
+        test_file_attribute_file_function_soldermask,
+        SolderMask,
+        "Soldermask"
+    );
+    test_position_and_index!(test_file_attribute_file_function_legend, Legend, "Legend");
+    test_layer_and_position!(
+        test_file_attribute_file_function_component,
+        Component,
+        "Component"
+    );
+    test_position!(test_file_attribute_file_function_paste, Paste, "Paste");
+    test_position!(test_file_attribute_file_function_glue, Glue, "Glue");
+    test_position_and_index!(
+        test_file_attribute_file_function_carbonmask,
+        CarbonMask,
+        "Carbonmask"
+    );
+    test_position_and_index!(
+        test_file_attribute_file_function_goldmask,
+        GoldMask,
+        "Goldmask"
+    );
+    test_position_and_index!(
+        test_file_attribute_file_function_heatsinkmask,
+        HeatsinkMask,
+        "Heatsinkmask"
+    );
+    test_position_and_index!(
+        test_file_attribute_file_function_peelablemask,
+        PeelableMask,
+        "Peelablemask"
+    );
+    test_position_and_index!(
+        test_file_attribute_file_function_silvermask,
+        SilverMask,
+        "Silvermask"
+    );
+    test_position_and_index!(
+        test_file_attribute_file_function_tinmask,
+        TinMask,
+        "Tinmask"
+    );
+    test_position!(
+        test_file_attribute_file_function_depthroute,
+        DepthRoute,
+        "Depthrout"
+    );
+    test_optional_position!(test_file_attribute_file_function_vcut, VCut, "Vcut");
+    test_simple!(
+        test_file_attribute_file_function_viafill,
+        ViaFill,
+        "Viafill"
+    );
+    test_position!(test_file_attribute_file_function_pads, Pads, "Pads");
+    test_string!(test_file_attribute_file_function_other, Other, "Other");
+    test_simple!(
+        test_file_attribute_file_function_drillmap,
+        DrillMap,
+        "Drillmap"
+    );
+    test_simple!(
+        test_file_attribute_file_function_fabricationdrawing,
+        FabricationDrawing,
+        "FabricationDrawing"
+    );
+    test_simple!(
+        test_file_attribute_file_function_vcutmap,
+        VCutMap,
+        "Vcutmap"
+    );
+    test_position!(
+        test_file_attribute_file_function_assemblydrawing,
+        AssemblyDrawing,
+        "AssemblyDrawing"
+    );
+    test_simple!(
+        test_file_attribute_file_function_arraydrawing,
+        ArrayDrawing,
+        "ArrayDrawing"
+    );
+    test_string!(
+        test_file_attribute_file_function_otherdrawing,
+        OtherDrawing,
+        "OtherDrawing"
+    );
+
+    #[test]
+    fn test_file_attribute_file_polarity() {
+        let pol = ExtendedCode::FileAttribute(FileAttribute::FilePolarity(FilePolarity::Positive));
+        assert_code!(pol, "%TF.FilePolarity,Positive*%\n");
+
+        let pol = ExtendedCode::FileAttribute(FileAttribute::FilePolarity(FilePolarity::Negative));
+        assert_code!(pol, "%TF.FilePolarity,Negative*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_file_function_md5() {
+        let md5 = ExtendedCode::FileAttribute(FileAttribute::Md5("abcd1234".into()));
+        assert_code!(md5, "%TF.MD5,abcd1234*%\n");
+    }
+
+    #[test]
+    fn test_file_attribute_file_function_user_defined() {
+        let owner = ExtendedCode::FileAttribute(FileAttribute::UserDefined {
+            name: "Authors".to_string(),
+            values: vec!["Author 1".to_string(), "Author 2".to_string()],
+        });
+        assert_code!(owner, "%TF.Authors,Author 1,Author 2*%\n");
+    }
+
+    #[test]
+    fn test_aperture_attribute() {
         // Test with Profile (found in "All data layers" section of the enum)
         let function = ExtendedCode::ApertureAttribute(ApertureAttribute::ApertureFunction(
             ApertureFunction::Profile,