+++ CHANGELOG.md
@@ -17,12 +17,15 @@ Possible log types:
 - [added] `MacroBoolean`.
 - [added] `MacroInteger`.
 - [added] Serialization of `ApertureAttribute` (previously it would panic on serializing them).
+- [added] Serialization of `FileAttribute` (previously it would panic on serializing them).
 - [changed] Replaced `Aperture::Other` with `Aperture::Macro`, the latter having option arguments.
 - [changed] Use `MacroBoolean` instead of `bool` for `exposure` on macro primitives.
 - [changed] `PolygonPrimitive::new` uses a `MacroInteger` for the `vertices` argument.
 - [changed] Replaced dependency on `conv::TryFrom` with `std::convert::TryFrom` (since TryFrom was stabilized in Rust 1.34).
 - [changed] Improved the API of `Coordinates`/`CoordinateOffsets` so that it's easier to build partial coordinates by 
   using a trait to convert from integer types or optional integer types.
+- [changed] `Scoring` replaced with `VCut` to match spec.
+- [changed] Various other changes to `ApertureAttribute`, `FileFunction` and `FileAttribute` to match spec.
 - [deprecated] `*Primitive::exposure_on`, replaced with `*Primitive::with_exposure`.
 
 ### v0.3.0 (2022-07-05)